+++ CHANGES.rst
@@ -15,6 +15,8 @@ Bug fixes
 * GH248 Fix an issue causing a ValueError to be raised when using `dask_index_on` on non-integer columns
 * GH255 Fix an issue causing the python interpreter to shut down when reading an
   empty file (see also https://issues.apache.org/jira/browse/ARROW-8142)
+* GH259 Fix an issue where Timestamp with timezone produce Schema violation
+
 
 Version 3.8.0 (2020-03-12)
 ==========================