+++ CHANGES.rst
@@ -13,6 +13,7 @@ Improvements
 Bug fixes
 ^^^^^^^^^
 * GH262: Raise an exception when trying to partition on a column with null values to prevent silent data loss
+* Fix multiple index creation issues (cutting data, crashing) for ``uint`` data
 
 
 Version 3.8.1 (2020-03-20)