+++ CHANGES.rst
@@ -16,6 +16,7 @@ Bug fixes
 * Fix multiple index creation issues (cutting data, crashing) for ``uint`` data
 * Fix index update issues for some types resulting in ``TypeError: Trying to update an index with different types...``
   messages.
+* Fix issues where index creation with empty partitions can lead to ``ValueError: Trying to create non-typesafe index``
 
 
 Version 3.8.1 (2020-03-20)